pyproject.toml
[tool.poetry]
name = "matrix-validator"
version = "0.0.0"
description = "matrix-validator"
authors = ["Nicolas Matentzoglu <nico@semanticly.org>"]
license = "BSD-3"
readme = "README.md"
packages = [
    { include = "matrix_validator", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "*"
polars-lts-cpu = "^1.23.0"
importlib-metadata = ">=8.2.0"
pandera = "^0.22.1"
tqdm = "*"
biolink-model = "4.2.6rc4"
patito = "0.8.3"
matrix-schema = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = {version = ">=8.3.2"}
tox = {version = ">=4.16.0"}
pre-commit = {version = ">=3.3.3"}
ruff = "^0.11.13"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = {version = ">=6.1.3"}
sphinx-rtd-theme = {version = ">=1.0.0"}
sphinx-autodoc-typehints = {version = ">=1.2.0"}
sphinx-click = {version = ">=4.3.0"}
myst-parser = {version = ">=0.18.1"}

[tool.poetry.scripts]
matrix-validator = "matrix_validator.cli:main"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.black]
line-length = 140
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
line_length = 140

[tool.ruff]
lint.extend-ignore = [
    "D211",  # `no-blank-line-before-class`
    "D212",  # `multi-line-summary-first-line`
    "D203",  # `blank-line-before-docstring`
    ]
line-length = 140

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]

# Select or ignore from https://beta.ruff.rs/docs/rules/
lint.select = [
    "B",  # bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle errors
    "F",  # Pyflakes
    "I",  # isort
    "S",  # flake8-bandit
    "W",  # Warning
]

lint.unfixable = []
target-version = "py310"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.codespell]
skip = "*.po,*.ts,.git,pyproject.toml"
count = ""
quiet-level = 3
# ignore-words-list = ""

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
